+++ test/candidate_test.dart
@@ -2,13 +2,30 @@ import 'package:test/test.dart';
 import 'package:wheatley/wheatley.dart';
 
 void main() {
-  late final Iterable<Candidate<int>> Function(int) shrinkInt;
-  shrinkInt = (s) => s > 0 ? [Candidate(s - 1, shrink: shrinkInt)] : [];
+  late final Iterable<int> Function(int) shrinkInt;
+  shrinkInt = (s) => s > 0 ? [s - 1] : [];
 
   group('Candidate', () {
     group('nullable', () {
       test('adds a null candidate value', () {
         expect(Candidate(1, shrink: shrinkInt).nullable.allValues, [1, 0, null]);
+      });
+    });
+    group('map', () {
+      test('should apply the mapper to the value', () {
+        final candidate1 = Candidate(4, shrink: shrinkInt);
+        final candidate2 = candidate1.map((value) => value * 2);
+
+        expect(candidate1.allValues.toList(), [4, 3, 2, 1, 0]);
+        expect(candidate2.allValues.toList(), [8, 6, 4, 2, 0]);
+      });
+      test('creates a candidate that disposes both the original and the mapped value', () {
+        final disposedValues = <int>[];
+        final candidate1 = Candidate(4, shrink: shrinkInt, dispose: disposedValues.add);
+        final candidate2 = candidate1.map((value) => value * 2, dispose: disposedValues.add);
+
+        candidate2.dispose();
+        expect(disposedValues, [4, 8]);
       });
     });
     group('flatMap', () {
@@ -17,8 +34,28 @@ void main() {
         final candidate2 = candidate1.flatMap((value) => Candidate(value * 2, shrink: shrinkInt));
 
         expect(candidate1.allValues.toList(), [4, 3, 2, 1, 0]);
-        // Steps down from the first shrunk value of candidate1 mapped with mapper, 3 * 2 = 6
-        expect(candidate2.allValues.toList(), [8, 6, 5, 4, 3, 2, 1, 0]);
+        expect(candidate2.allValues.toList(), [8, 7, 6, 5, 4, 3, 2, 1, 0]);
+      });
+      test('creates a candidate that disposes both the original and the mapped value', () {
+        final disposedValues = <int>[];
+        final candidate1 = Candidate(4, shrink: shrinkInt, dispose: disposedValues.add);
+        final candidate2 = candidate1.flatMap(
+          (value) => Candidate(value * 2, shrink: shrinkInt, dispose: disposedValues.add),
+        );
+
+        candidate2.dispose();
+        expect(disposedValues, [4, 8]);
+      });
+    });
+    group('zip', () {
+      test('creates a candidate that disposes both original candidates', () {
+        final disposedValues = <int>[];
+        final candidate1 = Candidate(4, shrink: shrinkInt, dispose: disposedValues.add);
+        final candidate2 = Candidate(5, shrink: shrinkInt, dispose: disposedValues.add);
+        final zippedCandidate = candidate1.zip(candidate2);
+
+        zippedCandidate.dispose();
+        expect(disposedValues, [4, 5]);
       });
     });
     group('shrinkUntilDone', () {
@@ -27,9 +64,6 @@ void main() {
       });
       test('calls dispose on each candidate except the last one', () async {
         final disposedValues = <int>[];
-        late final Iterable<Candidate<int>> Function(int) shrinkInt;
-        shrinkInt = (s) => s > 0 ? [Candidate(s - 1, shrink: shrinkInt, dispose: disposedValues.add)] : [];
-
         final candidate = Candidate(4, shrink: shrinkInt, dispose: disposedValues.add);
         await candidate.shrinkUntilDone((_) => throw Exception());
         expect(disposedValues, [4, 3, 2, 1]);
@@ -38,9 +72,6 @@ void main() {
         'calls dispose on each candidate except the last (null) one on a candidate from Candidate.nullable',
         () async {
           final disposedValues = <int>[];
-
-          late final Iterable<Candidate<int>> Function(int) shrinkInt;
-          shrinkInt = (s) => s > 0 ? [Candidate(s - 1, shrink: shrinkInt, dispose: disposedValues.add)] : [];
 
           final candidate = Candidate(4, shrink: shrinkInt, dispose: disposedValues.add).nullable;
           await candidate.shrinkUntilDone((v) => throw Exception());